Обновление проекта

+++ db/seed.js
@@ -2,7 +2,14 @@ const sequelize = require('./index');
 const Place = require('../models/Place');
 const { Route } = require('../models/Route');
 
+const forceReset = process.argv.includes('--force');
+
 async function seed() {
+    if (!forceReset) {
+        console.error('Сидирование остановлено: эта команда полностью очищает базу данных. Используй: node db/seed.js --force');
+        process.exit(1);
+    }
+
     await sequelize.sync({ force: true });
 
     const places = await Place.bulkCreate([
